+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# <a href='https://github.com/mrjackwills/obliqoro/releases/tag/v0.3.1'>v0.3.1</a>
+### 2025-08-28
+
+### Fixes
++ use tauri plugin for url & path opening, [6db5f88b](https://github.com/mrjackwills/obliqoro/commit/6db5f88b62c971c41050bd54120878809bb359ae)
+
+see [v0.3.0 release notes](https://github.com/mrjackwills/obliqoro/releases/tag/v0.3.0) for more information about v0.3 release
+
 # <a href='https://github.com/mrjackwills/obliqoro/releases/tag/v0.3.0'>v0.3.0</a>
 ### 2025-08-28